HomebrewFormula/dome-key.rb: Fix `:lion`/"10.7" `depends_on`
Turns out we didn't solve the problem with the previous commit, as
"10.7" still needs to be in Homebrew's `MacOS::Version::SYMBOLS` hash in
order to be recognised.

To work around this, depend on Lion if Homebrew provides it, otherwise
display a message in the Caveats section about the minimum OS version.